gr was spamming its temp, was an intr instaead of a long

+++ gr/gr.ino
@@ -48,7 +48,7 @@ long time_c_pos[]={0, 0, 0, 0, 0, 0, 0, 0};        //zeit der letzen steigenden 
 
 //constants and variables for t states (temperatur über dht22 an digitalem pin)
 const int num_t_states=2;
-const int period_t=1800000;                                                                                  //update periode in ms
+const long period_t=1800000;                                                                                  //update periode in ms
 const String t_address[]={"TI_GR", "TI_GR_A"};                                                                                                           //addresse
 const int t_pin[]={ 22, 24};
 int value_t[]={ 0,   0};                                            //temperatur
